% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_plot.R
\name{fun_plot}
\alias{fun_plot}
\title{plot sin(x) and sin(x/2) functions}
\usage{
fun_plot(from = -pi, to = pi, ...)
}
\arguments{
\item{from}{the starting point for x-axis}

\item{to}{the ending point for x-axis}

\item{...}{other parameters directly passed to curve()}
}
\value{
Null
}
\description{
calling curve() to plot these functions.
}
\examples{
fun_plot(-pi, pi)

}
